requirements.txt
pyyaml~=6.0.3
python-dotenv~=1.2.2
peewee~=4.0.4
fastapi~=0.135.3
uvicorn~=0.44.0
pydantic~=2.0
